fix: add lerp_ dtype cast in muon_step_fused for ROCm 7 compatibility
Apply explicit dtype casting to the second_momentum_buffer.lerp_()
call in muon_step_fused, matching the pattern used in adamw_step_fused.

This fixes compatibility with ROCm 7 (PyTorch 2.10) which has the same
strict dtype checking for lerp_ weight arguments.

Ref: PR review comment

+++ train.py
@@ -359,7 +359,7 @@ def muon_step_fused(stacked_grads, stacked_params, momentum_buffer, second_momen
     red_dim_size = g.size(red_dim)
     v_norm_sq = v_mean.sum(dim=(-2, -1), keepdim=True) * red_dim_size
     v_norm = v_norm_sq.sqrt()
-    second_momentum_buffer.lerp_(v_mean.to(dtype=second_momentum_buffer.dtype), 1 - beta2)
+    second_momentum_buffer.lerp_(v_mean.to(dtype=second_momentum_buffer.dtype), (1 - beta2).to(dtype=second_momentum_buffer.dtype))
     step_size = second_momentum_buffer.clamp_min(1e-10).rsqrt()
     scaled_sq_sum = (v_mean * red_dim_size) * step_size.float().square()
     v_norm_new = scaled_sq_sum.sum(dim=(-2, -1), keepdim=True).sqrt()
